% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/commuteDS.R
\name{commuteDS}
\alias{commuteDS}
\title{function to indicate when an individual is moving between two locations}
\usage{
commuteDS(input, id_col, loc1_col, loc2_col)
}
\arguments{
\item{input}{name of the object (usually SpatialPointsDataFrame) which contains the
points of the individuals' journeys and the two locations of interest}

\item{id_col}{name of the column in the data frame that identifies which points belong to
each individual}

\item{loc1_col}{name of the column in the data frame that indicates when the individual
is at the first location}

\item{loc2_col}{name of the column in the data frame that indicates when the individual
is at the second location}
}
\value{
a logical vector, with TRUE indicating that the individual was travelling
between the two locations
}
\description{
This function is used to determine whether an individual is moving between
two locations, for example when commuting between home and work.
}
\details{


Several assumptions are made about the nature of the data being examined.
 It is assumed that some groundwork has been done to provide a set of flags
 to indicate when the individual is at either of the locations. This data is grouped per
  individual, the groupings are defined by an identifier column. It is most likely that
  the input data will be a SpatialPointsDataFrame. The points making up an individual's
 journey should be temporally ordered. The function is able to detect when the individual
  leaves one location and arrives at the other and vice versa. Journeys that start and
  arrive at the same location are not counted.
}
\author{
Bishop, T.
}

