% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coordsToLinesDS.R
\name{coordsToLinesDS}
\alias{coordsToLinesDS}
\title{Convert groups of points to lines}
\usage{
coordsToLinesDS(coords, group)
}
\arguments{
\item{coords}{object of class SpatialPointsDataFrame to be converted
to a SpatialLinesDataFrame}

\item{group}{string indicating the name of the column in the data frame that defines
the grouping of the sets of points}
}
\value{
an object of class SpatialLines, each line having an ID taken from the column
specified by the group variable
}
\description{
This function converts a SpatialPointsDataFrame object into a
 SpatialLines object
}
\details{


The input object consists of sets of points grouped by an identifier. The
function takes each group of points and converts them into a line with an associated
identifier
}
\author{
Bishop, T.
}

