% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoSummaryDS.R
\name{geoSummaryDS}
\alias{geoSummaryDS}
\title{Summary for geo objects}
\usage{
geoSummaryDS(input)
}
\arguments{
\item{input}{sp object as defined in package sp}
}
\value{
SpatialPolygons (or a SpatialPolygonsDataFrame if by_id=TRUE and
input has a data.frame); if negative width(s) lead the object to disappear,
NULL is returned for by_id FALSE, and component Polygons objects are dropped if
empty for byid TRUE; the SpatialPolygonsDataFrame is subsetted by row.names or id
if given to retain non-empty geometry rows
}
\description{
This function wraps summary
}
\details{
See the gBuffer() function from rgeos package for more details
}
\author{
Bishop, T.
}
