% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapplyDS.R
\name{lapplyDS}
\alias{lapplyDS}
\title{Wrapper for \code{lapply()} function from sp and rgeos package}
\usage{
lapplyDS(input, fun = NULL)
}
\arguments{
\item{input}{a list}

\item{fun}{a function; see value}
}
\value{
as per lapply
}
\description{
This function is a wrapper for the \code{lapply()} function
}
\details{
See the lapply() function
}
\author{
Bishop, T.
}
