% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nothingDS.R
\name{nothingDS}
\alias{nothingDS}
\title{Does nothing}
\usage{
nothingDS(x)
}
\arguments{
\item{x}{a thing}
}
\value{
a message
}
\description{
This function does nothing
}
\details{
it is still nothing
}
\author{
Bishop, T.
}
