% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overDS.R
\name{overDS}
\alias{overDS}
\title{Wrapper for \code{over()} function from sp and rgeos package}
\usage{
overDS(x, y)
}
\arguments{
\item{x}{geometry (locations) of the queries}

\item{y}{layer from which the geometries or attributes are queried}
}
\value{
A vector of length length(x) with the (first) index of y for each geometry
(point, grid cell centre, polygon or lines) in x. If y has attribute data, attribute
data are returned in a data.frame with number of rows equal to length(x)
}
\description{
This function is a wrapper for the \code{over()} function from the
sp and rgeos package.consistent spatial overlay for points, grids and polygons:
at the spatial locations of object x retrieves the indexes or attributes from
spatial object y. \code{over()} can be seen as a left outer join in SQL;
the match is a spatial intersection.
}
\details{


See the over() function from sp and rgeos package for more details
}
\author{
Bishop, T.
}

