% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/proj4stringDS.R
\name{proj4stringDS}
\alias{proj4stringDS}
\title{Wrapper for proj4string() function from sp and rgdal packages}
\usage{
proj4stringDS(x, projStr)
}
\arguments{
\item{x}{object of class "spatial-class"}

\item{projStr}{a valid proj4 epsg coordinate system identifier as a string e.g. 29902 for
Ireland}
}
\value{
object of class "spatial-class" with the proj4string set to the value passed
to the function
}
\description{
This function is a wrapper for the proj4string() function from the
sp and rgdal packages
}
\details{


See the proj4string() function from sp package for more details
}
\author{
Bishop, T.
}

