% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpatialLinesDataFrameDS.R
\name{SpatialLinesDataFrameDS}
\alias{SpatialLinesDataFrameDS}
\title{Wrapper for SpatialLinesDataFrame() function from sp package}
\usage{
SpatialLinesDataFrameDS(lines, data)
}
\arguments{
\item{lines}{object of class SpatialLines-class}

\item{data}{object of class data.frame; the number of rows in data should equal
the number of Lines elements in lines.}
}
\value{
object of class SpatialLinesDataFrame
}
\description{
This function is a wrapper for the SpatialLinesDataFrame() function from the
sp package
}
\details{


See the SpatialLinesDataFrame() function from sp package for more details.
create objects of class SpatialLinesDataFrame from lists of Lines objects
and data.frames
}
\author{
Bishop, T.
}

