% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTransformDS.R
\name{spTransformDS}
\alias{spTransformDS}
\title{Wrapper for spTransform() function from sp and rgdal packages}
\usage{
spTransformDS(input, projStr)
}
\arguments{
\item{input}{object of class "spatial-class"}

\item{projStr}{a valid proj4 epsg coordinate system identifier as a string e.g. 29902 for
Ireland}
}
\value{
object of class "spatial-class" with the coordinate system transformed to that
specified in projStr
}
\description{
This function is a wrapper for the spTransform() function from the
sp and rgdal packages
}
\details{
See the spTransform() function from sp package for more details
}
\author{
Bishop, T.
}
